# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/{userId}/games": {
      "put": {
        "operationId": "UsersController_updateUserGames",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateGamesBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The games have been successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Modify games owned by user",
        "tags": [
          "users"
        ]
      }
    },
...
    "/collection/users/{userId}/games": {
      "get": {
        "deprecated": false,
        "operationId": "CollectionController_getGamesOwnedByUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all games owned by the user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameCompleteDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all games owned by a user",
        "tags": [
          "collection"
        ]
      }
    },
    "/collection/users/{userId}/games/{gameId}": {
      "get": {
        "deprecated": false,
        "operationId": "CollectionController_getGameViewByUserId",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "gameId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Game view for the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameViewDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Game view for a user (not owned necessarily)",
        "tags": [
          "collection"
        ]
      },
      "post": {
        "deprecated": false,
        "operationId": "CollectionController_addGameToUserCollection",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "gameId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Game added to the user's collection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a game to a user's collection",
        "tags": [
          "collection"
        ]
      },
      "delete": {
        "deprecated": false,
        "operationId": "CollectionController_removeGameFromUserCollection",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "gameId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Game removed from the user's collection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a game from a user's collection",
        "tags": [
          "collection"
        ]
      }
    },
    "/collection/users/{userId}/games/{gameId}/ownership": {
      "patch": {
        "deprecated": false,
        "operationId": "CollectionController_updateGameOwnership",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "gameId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGameOwnedDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated game ownership details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameOwnedDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update ownership details of a game",
        "tags": [
          "collection"
        ]
      }
    },
    "/collection/users/{userId}/games/{gameId}/wishlist": {
      "put": {
        "deprecated": false,
        "operationId": "CollectionController_toggleWishlist",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "gameId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WishlistResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Toggle the wishlist status of the game",
        "tags": [
          "collection"
        ]
      }
    },
    "/collection/users/{userId}/browse/games": {
      "get": {
        "deprecated": false,
        "operationId": "CollectionController_getGamesNotOwnedByUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Title search across supported translations.",
            "schema": {
              "example": "catan",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 20,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all games not owned by the user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BrowseGamesResultDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Browse games not owned by a user",
        "tags": [
          "collection"
        ]
      }
    },
    "/collection/users/{userId}/reviews": {
      "get": {
        "deprecated": false,
        "operationId": "CollectionController_getReviewsOfUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all reviews of the user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameReviewWithGameDataDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all reviews of a user",
        "tags": [
          "collection"
        ]
      }
    },
    "/collection/users/{userId}/reviews/{gameId}": {
      "post": {
        "deprecated": false,
        "operationId": "CollectionController_saveGameReview",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "gameId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameReviewBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The review has been succesfully saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Save a game review",
        "tags": [
          "collection"
        ]
      }
    },
    "/collection/users/{userId}/wishlist": {
      "get": {
        "deprecated": false,
        "operationId": "CollectionController_getUserWishlist",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all wishlist of the user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all wishlist of a user",
        "tags": [
          "collection"
        ]
      }
    },
    "/collection/users/{userId}/recent-activity": {
      "get": {
        "deprecated": false,
        "operationId": "CollectionController_getUserCollectionActivities",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all recent activity of the user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CollectionActivityDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all recent activity of a user's collection",
        "tags": [
          "collection"
        ]
      }
    },
...
  },
  "info": {
    "title": "Board Vault API",
    "description": "Versioned contract snapshot for the Board Vault social group workspace API.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UserUpdateGamesBody": {
        "type": "object",
        "properties": {
          "gamesToAdd": {
            "description": "The games to add to user",
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "gamesToRemove": {
            "description": "The games to remove from user",
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "gamesToAdd",
          "gamesToRemove"
        ]
      },
      "SuccessDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true,
            "description": "The success status of the operation."
          }
        },
        "required": [
          "success"
        ]
      },
...
      "GameCompleteDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 12345
          },
          "imageUrl": {
            "type": "string",
            "example": "https://www.example.com/image.jpg"
          },
          "gameAvgDuration": {
            "type": "number",
            "example": 120,
            "description": "The average duration of the game in minutes."
          },
          "minPlayers": {
            "type": "number",
            "example": 3,
            "description": "The minimum number of players required to play the game."
          },
          "maxPlayers": {
            "type": "number",
            "example": 4,
            "description": "The maximum number of players that can play the game."
          },
          "title": {
            "type": "string",
            "example": "Game Title"
          },
          "titleTranslations": {
            "type": "object",
            "example": {
              "en": "Game Title"
            }
          }
        },
        "required": [
          "id",
          "imageUrl",
          "gameAvgDuration",
          "minPlayers",
          "maxPlayers",
          "titleTranslations"
        ]
      },
...
      "GameDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 12345
          },
          "imageUrl": {
            "type": "string",
            "example": "https://www.example.com/image.jpg"
          },
          "gameAvgDuration": {
            "type": "number",
            "example": 120,
            "description": "The average duration of the game in minutes."
          },
          "minPlayers": {
            "type": "number",
            "example": 3,
            "description": "The minimum number of players required to play the game."
          },
          "maxPlayers": {
            "type": "number",
            "example": 4,
            "description": "The maximum number of players that can play the game."
          }
        },
        "required": [
          "id",
          "imageUrl",
          "gameAvgDuration",
          "minPlayers",
          "maxPlayers"
        ]
      },
...
      "BrowseGamesPaginationDto": {
        "type": "object",
        "properties": {
          "currentPage": {
            "type": "number",
            "example": 1
          },
          "totalPages": {
            "type": "number",
            "example": 10
          },
          "totalItems": {
            "type": "number",
            "example": 100
          },
          "itemsPerPage": {
            "type": "number",
            "example": 10
          }
        },
        "required": [
          "currentPage",
          "totalPages",
          "totalItems",
          "itemsPerPage"
        ]
      },
...
      "GameOwnershipDto": {
        "type": "object",
        "properties": {
          "purchaseDate": {
            "type": "object",
            "example": "2026-09-01",
            "nullable": true
          },
          "purchasePrice": {
            "type": "object",
            "example": 42.5,
            "nullable": true
          },
          "purchaseNotes": {
            "type": "object",
            "example": "Bought for the group",
            "nullable": true
          }
        },
        "required": [
          "purchaseDate",
          "purchasePrice",
          "purchaseNotes"
        ]
      },
      "GameViewTagDto": {
        "type": "object",
        "properties": {
          "tag": {
            "type": "string",
            "example": "Strategy"
          },
          "category": {
            "type": "string",
            "example": "Genre"
          }
        },
        "required": [
          "tag",
          "category"
        ]
      },
      "GameWishlistDto": {
        "type": "object",
        "properties": {
          "dateAdded": {
            "type": "string",
            "example": "2026-09-01"
          },
          "notes": {
            "type": "string",
            "example": "Suggested by a friend"
          }
        },
        "required": [
          "dateAdded",
          "notes"
        ]
      },
      "GameViewRatingSummaryDto": {
        "type": "object",
        "properties": {
          "review": {
            "type": "number",
            "example": 8.5
          },
          "count": {
            "type": "number",
            "example": 4
          }
        },
        "required": [
          "review",
          "count"
        ]
      },
      "GameViewRatingDto": {
        "type": "object",
        "properties": {
          "userRating": {
            "type": "object",
            "example": 9,
            "nullable": true
          },
          "avgGroupsRating": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/GameViewRatingSummaryDto"
              }
            ]
          },
          "avgGlobalRating": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/GameViewRatingSummaryDto"
              }
            ]
          }
        },
        "required": [
          "userRating",
          "avgGroupsRating",
          "avgGlobalRating"
        ]
      },
      "GameViewDto": {
        "type": "object",
        "properties": {
          "gameData": {
            "$ref": "#/components/schemas/GameCompleteDto"
          },
          "ownedGameData": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/GameOwnershipDto"
              }
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameViewTagDto"
            }
          },
          "wishlistedGameData": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/GameWishlistDto"
              }
            ]
          },
          "ratingData": {
            "$ref": "#/components/schemas/GameViewRatingDto"
          },
          "similarGames": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameCompleteDto"
            }
          }
        },
        "required": [
          "gameData",
          "ownedGameData",
          "tags",
          "wishlistedGameData",
          "ratingData",
          "similarGames"
        ]
      },
      "UpdateGameOwnedDto": {
        "type": "object",
        "properties": {
          "purchasePrice": {
            "type": "object",
            "example": 12345,
            "description": "The price of the game when purchased."
          },
          "purchaseDate": {
            "type": "object",
            "example": "2021-01-01",
            "description": "The date the game was purchased."
          },
          "purchaseNotes": {
            "type": "object",
            "example": "I got this as a gift from my friend.",
            "description": "Any notes about the purchase."
          }
        }
      },
      "GameOwnedDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "number",
            "example": 12345,
            "description": "The unique identifier for the Account."
          },
          "gameId": {
            "type": "number",
            "example": 12345,
            "description": "The unique identifier for the Game."
          },
          "purchasePrice": {
            "type": "object",
            "example": 12345,
            "description": "The price of the game when purchased."
          },
          "purchaseDate": {
            "type": "object",
            "example": "2021-01-01",
            "description": "The date the game was purchased."
          },
          "purchaseNotes": {
            "type": "object",
            "example": "I got this as a gift from my friend.",
            "description": "Any notes about the purchase."
          }
        },
        "required": [
          "accountId",
          "gameId",
          "purchasePrice",
          "purchaseDate",
          "purchaseNotes"
        ]
      },
      "WishlistResponseDto": {
        "type": "object",
        "properties": {
          "isWishlisted": {
            "type": "boolean",
            "example": true,
            "description": "Whether the game is wishlisted."
          }
        },
        "required": [
          "isWishlisted"
        ]
      },
      "BrowseGamesResultDto": {
        "type": "object",
        "properties": {
          "games": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameCompleteDto"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/BrowseGamesPaginationDto"
          }
        },
        "required": [
          "games",
          "pagination"
        ]
      },
      "GameReviewWithGameDataDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "number",
            "example": 12345
          },
          "gameId": {
            "type": "number",
            "example": 12345
          },
          "review": {
            "type": "number",
            "example": 8
          },
          "reviewDate": {
            "type": "string",
            "example": "2022-03-07T16:00:00.000Z"
          },
          "gameData": {
            "$ref": "#/components/schemas/GameCompleteDto"
          }
        },
        "required": [
          "accountId",
          "gameId",
          "review",
          "reviewDate",
          "gameData"
        ]
      },
      "CreateGameReviewBody": {
        "type": "object",
        "properties": {
          "review": {
            "type": "number",
            "example": 8,
            "minimum": 0,
            "maximum": 10
          }
        },
        "required": [
          "review"
        ]
      },
      "CollectionActivityDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 12345
          },
          "accountId": {
            "type": "number",
            "example": 12345
          },
          "gameId": {
            "type": "number",
            "example": 567
          },
          "actionType": {
            "type": "string",
            "example": "rated",
            "enum": [
              "added",
              "rated",
              "wishlisted",
              "unwishlisted",
              "updated",
              "removed"
            ]
          },
          "actionDetails": {
            "type": "object",
            "example": "{\"rating\": 4.5}",
            "description": "JSON string with additional details about the action"
          },
          "createdAt": {
            "type": "string",
            "example": "2024-09-28 10:02:39"
          }
        },
        "required": [
          "id",
          "accountId",
          "gameId",
          "actionType",
          "actionDetails",
          "createdAt"
        ]
      },
...
    }
  }
}